Move more logic into table grid component
